begin moving business logic out of VC
